+++ ofrak_core/src/ofrak/core/cramfs.py
@@ -1,5 +1,6 @@
 import asyncio
-import tempfile
+import os
+import tempfile312 as tempfile
 from dataclasses import dataclass
 from subprocess import CalledProcessError
 
@@ -14,17 +15,23 @@ from ofrak_type.range import Range
 
 # mkfs.cramfs from util-linux 2.38.1
 FSCK_CRAMFS = ComponentExternalTool(
-    "fsck.cramfs", "", "--help",
+    "fsck.cramfs",
+    "",
+    "--help",
 )
 MKFS_CRAMFS = ComponentExternalTool(
-    "mkfs.cramfs", "", "--help",
+    "mkfs.cramfs",
+    "",
+    "--help",
 )
+
 
 @dataclass
 class Cramfs(GenericBinary, FilesystemRoot):
     """
     Filesystem stored in CramFS format.
     """
+
 
 class CramfsUnpacker(Unpacker[None]):
     """
@@ -33,26 +40,26 @@ class CramfsUnpacker(Unpacker[None]):
 
     targets = (Cramfs,)
     children = (File, Folder, SpecialFileType)
-    external_dependnecies = (FSCK_CRAMFS,)
+    external_dependencies = (FSCK_CRAMFS,)
 
     async def unpack(self, resource: Resource, config=None):
         async with resource.temp_to_disk() as temp_path:
-            with tempfile.TemporaryDirectory() as temp_flush_dir:
-                # TODO
+            with tempfile.TemporaryDirectory() as temp_parent:
+                temp_flush_dir = os.path.join(temp_parent, "cramfs_extract")
                 cmd = [
                     FSCK_CRAMFS.tool,
                     f"--extract={temp_flush_dir}",
-                    temp_path
+                    temp_path,
                 ]
-            proc = await asyncio.create_subprocess_exec(
-                *cmd,
-            )
-            returncode = await proc.wait()
-            if proc.returncode:
-                raise CalledProcessError(returncode=returncode, cmd=cmd)
+                proc = await asyncio.create_subprocess_exec(
+                    *cmd,
+                )
+                returncode = await proc.wait()
+                if proc.returncode:
+                    raise CalledProcessError(returncode=returncode, cmd=cmd)
 
-            cramfs_view = await resource.view_as(Cramfs)
-            await cramfs_view.initialize_from_disk(temp_flush_dir)
+                cramfs_view = await resource.view_as(Cramfs)
+                await cramfs_view.initialize_from_disk(temp_flush_dir)
 
 
 class CramFSPacker(Packer[None]):
@@ -71,7 +78,7 @@ class CramFSPacker(Packer[None]):
             cmd = [
                 MKFS_CRAMFS.tool,
                 temp_flush_dir,
-                temp
+                temp.name,
             ]
             proc = await asyncio.create_subprocess_exec(
                 *cmd,
@@ -83,6 +90,7 @@ class CramFSPacker(Packer[None]):
                 new_data = new_fh.read()
             # Passing in the original range effectively replaces the original data with the new data
             resource.queue_patch(Range(0, await resource.get_data_length()), new_data)
+
 
 MagicDescriptionPattern.register(
     Cramfs, lambda s: s.startswith("Linux Compressed ROM File System data")